requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
playwright>=1.42.0
openai>=1.30.0
aiosqlite>=0.20.0
python-multipart>=0.0.9
httpx>=0.27.0
pydantic>=2.6.0
python-docx>=1.1.0
